Prevent output stream from hanging
- (fix) ensure output stream doesn't hang while calculating etag

The issue is that the transform stream being used for the new output
has a highwater mark to limit the maximum amount of data to buffer
while waiting for the reader to consume the output. Because the etag
code doesn't complete until we either finish reading the source stream
or hit the size limit, we get into a situation where we hang
indefinitely waiting for the transform stream to drain which will never
happen because we haven't yet piped it to the output. Setting the
highwater mark to the same thing as the size limit should make sure we
are allowed to buffer enough to allow the etag logic to complete.